+++ src/components/ha-date-range-picker.ts
@@ -14,7 +14,7 @@ import {
 import { customElement, property } from "lit/decorators";
 import { formatDateTime } from "../common/datetime/format_date_time";
 import { useAmPm } from "../common/datetime/use_am_pm";
-import { useFirstWeekdayIndex } from "../common/datetime/weekday";
+import { useFirstWeekdayIndex } from "../common/datetime/use_first_weekday";
 import { computeRTLDirection } from "../common/util/compute_rtl";
 import { HomeAssistant } from "../types";
 import "./date-range-picker";
