+++ src/panels/config/helpers/forms/ha-schedule-form.ts
@@ -19,7 +19,7 @@ import {
 import { customElement, property, state } from "lit/decorators";
 import { formatTime24h } from "../../../../common/datetime/format_time";
 import { useAmPm } from "../../../../common/datetime/use_am_pm";
-import { useFirstWeekdayIndex } from "../../../../common/datetime/weekday";
+import { useFirstWeekdayIndex } from "../../../../common/datetime/use_first_weekday";
 import { fireEvent } from "../../../../common/dom/fire_event";
 import "../../../../components/ha-icon-picker";
 import "../../../../components/ha-textfield";
